requirements.txt
adafruit-ampy==1.1.0
bitstring==3.1.9
black==21.12b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
esptool==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
pyserial==3.5
python-dotenv==0.19.2
reedsolo==1.5.4
six==1.16.0
tomli==1.2.3
typing-extensions==4.0.1
